requirements.txt
Pillow
requests
openai
python-dotenv
PyYAML
google
google-genai
pandas
openpyxl
